Update sqa checking script for other apps
